Move failing top-level components tests to failing_tests directory
There also appear to be some broken imports caused by something in the Dashboard.

This commit does not handle subcomponents, which will come in the next one.

The test summary is now:
```
Test Suites: 13 failed, 216 passed, 229 total
Tests:       43 failed, 11 skipped, 1360 passed, 1414 total
Snapshots:   620 passed, 620 total
Time:        23.552s
Ran all test suites.
```
