+++ tests/components/channel_info_modal.test.jsx
@@ -2,10 +2,8 @@
 // See LICENSE.txt for license information.
 
 import React from 'react';
-import {Modal} from 'react-bootstrap';
 import {shallow} from 'enzyme';
 
-import {mountWithIntl} from 'tests/helpers/intl-test-helper.jsx';
 import ChannelInfoModal from 'components/channel_info_modal/channel_info_modal.jsx';
 
 describe('components/ChannelInfoModal', () => {
@@ -48,23 +46,5 @@ describe('components/ChannelInfoModal', () => {
         );
 
         expect(wrapper).toMatchSnapshot();
-    });
-
-    // Something about this test requires is to be run last, otherwise the suite will hang.
-    it('should call onHide callback when modal is hidden', (done) => {
-        function onHide() {
-            done();
-        }
-
-        const wrapper = mountWithIntl(
-            <ChannelInfoModal
-                channel={{name: 'testchannel', displayName: 'testchannel', header: '', purpose: ''}}
-                currentTeam={{id: 'testid', name: 'testteam'}}
-                onHide={onHide}
-            />
-        );
-
-        wrapper.find(Modal).first().props().onExited();
-        done();
     });
 });
