+++ tests/components/edit_post_modal.test.jsx
@@ -4,7 +4,7 @@ import React from 'react';
 import {shallow} from 'enzyme';
 import ReactRouterEnzymeContext from 'react-router-enzyme-context';
 
-import {mountWithIntl} from 'tests/helpers/intl-test-helper.jsx';
+//import {mountWithIntl} from 'tests/helpers/intl-test-helper.jsx';
 import {Constants, ModalIdentifiers} from 'utils/constants';
 import DeletePostModal from 'components/delete_post_modal';
 import EditPostModal from 'components/edit_post_modal/edit_post_modal.jsx';
@@ -173,53 +173,6 @@ describe('components/EditPostModal', () => {
         expect(wrapper.state().showEmojiPicker).toBe(false);
     });
 
-    it('should add emoji to editText when an emoji is clicked', () => {
-        const options = new ReactRouterEnzymeContext();
-        const wrapper = mountWithIntl(createEditPost(), options.get());
-        wrapper.setState({editText: ''});
-        wrapper.instance().handleEmojiClick(null);
-        wrapper.instance().handleEmojiClick({});
-        wrapper.instance().handleEmojiClick({aliases: []});
-        expect(wrapper.state().editText).toBe('');
-
-        wrapper.setState({editText: ''});
-        wrapper.instance().handleEmojiClick({name: '+1', aliases: ['thumbsup']});
-        expect(wrapper.state().editText).toBe(':+1: ');
-
-        wrapper.setState({editText: 'test'});
-        wrapper.instance().handleEmojiClick({name: '-1', aliases: ['thumbsdown']});
-        expect(wrapper.state().editText).toBe('test :-1: ');
-
-        wrapper.setState({editText: 'test '});
-        wrapper.instance().handleEmojiClick({name: '-1', aliases: ['thumbsdown']});
-        expect(wrapper.state().editText).toBe('test :-1: ');
-    });
-
-    it('should set the focus and recalculate the size of the edit box after entering', () => {
-        const options = new ReactRouterEnzymeContext();
-        const wrapper = mountWithIntl(createEditPost(), options.get());
-        const instance = wrapper.instance();
-        const ref = instance.editbox;
-        ref.focus = jest.fn();
-        ref.recalculateSize = jest.fn();
-        expect(ref.focus).not.toBeCalled();
-        expect(ref.recalculateSize).not.toBeCalled();
-        instance.handleEntered();
-        expect(ref.focus).toBeCalled();
-        expect(ref.recalculateSize).toBeCalled();
-    });
-
-    it('should hide the preview when exiting', () => {
-        const options = new ReactRouterEnzymeContext();
-        const wrapper = mountWithIntl(createEditPost(), options.get());
-        const instance = wrapper.instance();
-        const ref = instance.editbox;
-        ref.hidePreview = jest.fn();
-        expect(ref.hidePreview).not.toBeCalled();
-        instance.handleExit();
-        expect(ref.hidePreview).toBeCalled();
-    });
-
     it('should close without saving when post text is not changed', () => {
         const actions = {
             editPost: jest.fn(),
@@ -375,56 +328,6 @@ describe('components/EditPostModal', () => {
         expect(instance.handleEdit).not.toBeCalled();
         instance.handleKeyDown({key: Constants.KeyCodes.ENTER[0], keyCode: Constants.KeyCodes.ENTER[1], ctrlKey: true});
         expect(instance.handleEdit).not.toBeCalled();
-    });
-
-    it('should handle edition on key press enter depending on the conditions', () => {
-        const options = new ReactRouterEnzymeContext();
-        global.navigator = {userAgent: 'Android'};
-        var wrapper = mountWithIntl(createEditPost({ctrlSend: true}), options.get());
-        var instance = wrapper.instance();
-        const preventDefault = jest.fn();
-        instance.handleEdit = jest.fn();
-        instance.handleEditKeyPress({which: 1, ctrlKey: true, preventDefault, shiftKey: false, altKey: false});
-        expect(instance.handleEdit).not.toBeCalled();
-        expect(preventDefault).not.toBeCalled();
-        instance.handleEditKeyPress({key: Constants.KeyCodes.ENTER[0], which: Constants.KeyCodes.ENTER[1], ctrlKey: false, preventDefault, shiftKey: false, altKey: false});
-        expect(instance.handleEdit).not.toBeCalled();
-        expect(preventDefault).not.toBeCalled();
-        instance.handleEditKeyPress({key: Constants.KeyCodes.ENTER[0], which: Constants.KeyCodes.ENTER[1], ctrlKey: true, preventDefault, shiftKey: false, altKey: false});
-        expect(instance.handleEdit).toBeCalled();
-        expect(preventDefault).toBeCalled();
-
-        global.navigator = {userAgent: 'Chrome'};
-        wrapper = mountWithIntl(createEditPost({ctrlSend: false}), options.get());
-        instance = wrapper.instance();
-        preventDefault.mockClear();
-        instance.handleEdit = jest.fn();
-        instance.handleEditKeyPress({which: 1, ctrlKey: true, preventDefault, shiftKey: false, altKey: false});
-        expect(instance.handleEdit).not.toBeCalled();
-        expect(preventDefault).not.toBeCalled();
-        instance.handleEditKeyPress({key: Constants.KeyCodes.ENTER[0], which: Constants.KeyCodes.ENTER[1], ctrlKey: true, preventDefault, shiftKey: true, altKey: false});
-        expect(instance.handleEdit).not.toBeCalled();
-        expect(preventDefault).not.toBeCalled();
-        instance.handleEditKeyPress({key: Constants.KeyCodes.ENTER[0], which: Constants.KeyCodes.ENTER[1], ctrlKey: true, preventDefault, shiftKey: false, altKey: true});
-        expect(instance.handleEdit).not.toBeCalled();
-        expect(preventDefault).not.toBeCalled();
-        instance.handleEditKeyPress({key: Constants.KeyCodes.ENTER[0], ctrlKey: true, preventDefault, shiftKey: false, altKey: false});
-        expect(instance.handleEdit).toBeCalled();
-        expect(preventDefault).toBeCalled();
-    });
-
-    it('should handle the escape key manually to hide the modal', () => {
-        const options = new ReactRouterEnzymeContext();
-        const wrapper = mountWithIntl(createEditPost({ctrlSend: true}), options.get());
-        const instance = wrapper.instance();
-        instance.handleHide = jest.fn();
-        instance.handleExit = jest.fn();
-
-        instance.handleKeyDown({keyCode: 1});
-        expect(instance.handleHide).not.toBeCalled();
-
-        instance.handleKeyDown({key: Constants.KeyCodes.ESCAPE[0], keyCode: Constants.KeyCodes.ESCAPE[1]});
-        expect(instance.handleHide).toBeCalled();
     });
 
     it('should handle the escape key manually to hide the modal, unless the emoji picker is shown', () => {
