+++ tests/components/failing_tests/channel_header.test.jsx
