+++ tests/components/failing_tests/channel_info_modal.test.jsx
@@ -1,0 +1,28 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+import React from 'react';
+import {Modal} from 'react-bootstrap';
+
+import {mountWithIntl} from 'tests/helpers/intl-test-helper.jsx';
+import ChannelInfoModal from 'components/channel_info_modal/channel_info_modal.jsx';
+
+describe('components/ChannelInfoModal', () => {
+    // Something about this test requires is to be run last, otherwise the suite will hang.
+    it('should call onHide callback when modal is hidden', (done) => {
+        function onHide() {
+            done();
+        }
+
+        const wrapper = mountWithIntl(
+            <ChannelInfoModal
+                channel={{name: 'testchannel', displayName: 'testchannel', header: '', purpose: ''}}
+                currentTeam={{id: 'testid', name: 'testteam'}}
+                onHide={onHide}
+            />
+        );
+
+        wrapper.find(Modal).first().props().onExited();
+        done();
+    });
+});
