+++ tests/components/failing_tests/toggle_modal_button_redux.test.jsx
