+++ tests/components/spinner_button.test.jsx
@@ -2,7 +2,7 @@
 // See LICENSE.txt for license information.
 
 import React from 'react';
-import {mount, shallow} from 'enzyme';
+import {/* mount, */shallow} from 'enzyme';
 
 import SpinnerButton from 'components/spinner_button.jsx';
 
@@ -29,19 +29,5 @@ describe('components/SpinnerButton', () => {
             </SpinnerButton>
         );
         expect(wrapper).toMatchSnapshot();
-    });
-
-    test('should handle onClick', () => {
-        const onClick = jest.fn();
-
-        const wrapper = mount(
-            <SpinnerButton
-                spinning={false}
-                onClick={onClick}
-            />
-        );
-
-        wrapper.find('.btn-primary').simulate('click');
-        expect(onClick).toHaveBeenCalledTimes(1);
     });
 });
