+++ tests/components/system_notice.test.jsx
@@ -26,11 +26,6 @@ describe('components/SystemNotice', () => {
         },
     };
 
-    test('should match snapshot for regular user, regular notice', () => {
-        const wrapper = shallow(<SystemNotice {...baseProps}/>);
-        expect(wrapper).toMatchSnapshot();
-    });
-
     test('should match snapshot for regular user, no notice', () => {
         const props = {...baseProps, notices: []};
         const wrapper = shallow(<SystemNotice {...props}/>);
@@ -39,24 +34,6 @@ describe('components/SystemNotice', () => {
 
     test('should match snapshot for regular user, admin notice', () => {
         const props = {...baseProps, notices: [{name: 'notice1', adminOnly: true, title: 'some title', icon: mattermostIcon, body: 'some body', allowForget: true, show: () => true}]};
-        const wrapper = shallow(<SystemNotice {...props}/>);
-        expect(wrapper).toMatchSnapshot();
-    });
-
-    test('should match snapshot for regular user, admin and regular notice', () => {
-        const props = {...baseProps, notices: [{name: 'notice1', adminOnly: true, title: 'some title', icon: mattermostIcon, body: 'some body', allowForget: true}, {name: 'notice2', adminOnly: false, title: 'some title2', icon: mattermostIcon, body: 'some body2', allowForget: true, show: () => true}]};
-        const wrapper = shallow(<SystemNotice {...props}/>);
-        expect(wrapper).toMatchSnapshot();
-    });
-
-    test('should match snapshot for admin, regular notice', () => {
-        const props = {...baseProps, isSystemAdmin: true};
-        const wrapper = shallow(<SystemNotice {...props}/>);
-        expect(wrapper).toMatchSnapshot();
-    });
-
-    test('should match snapshot for admin, admin notice', () => {
-        const props = {...baseProps, isSystemAdmin: true, notices: [{name: 'notice1', adminOnly: true, title: 'some title', icon: mattermostIcon, body: 'some body', allowForget: true, show: () => true}]};
         const wrapper = shallow(<SystemNotice {...props}/>);
         expect(wrapper).toMatchSnapshot();
     });
@@ -75,18 +52,6 @@ describe('components/SystemNotice', () => {
 
     test('should match snapshot for show function returning false', () => {
         const props = {...baseProps, notices: [{name: 'notice1', adminOnly: false, title: 'some title', icon: mattermostIcon, body: 'some body', allowForget: true, show: () => false}]};
-        const wrapper = shallow(<SystemNotice {...props}/>);
-        expect(wrapper).toMatchSnapshot();
-    });
-
-    test('should match snapshot for show function returning true', () => {
-        const props = {...baseProps, notices: [{name: 'notice1', adminOnly: false, title: 'some title', icon: mattermostIcon, body: 'some body', allowForget: true, show: () => true}]};
-        const wrapper = shallow(<SystemNotice {...props}/>);
-        expect(wrapper).toMatchSnapshot();
-    });
-
-    test('should match snapshot for with allowForget equal false', () => {
-        const props = {...baseProps, notices: [{name: 'notice1', adminOnly: false, title: 'some title', icon: mattermostIcon, body: 'some body', allowForget: false, show: () => true}]};
         const wrapper = shallow(<SystemNotice {...props}/>);
         expect(wrapper).toMatchSnapshot();
     });
